+++ src/main/java/com/example/webscraper/WebScraperApplication.java
@@ -1,4 +1,4 @@
-package webscraper;
+package com.example.webscraper;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
